+++ examples/example.js
@@ -6,7 +6,8 @@ const exampleObj = {
   goodKey: 'I exist :)',
   nestedKey: {
     nestedGoodKey: 'Me too!'
-  }
+  },
+  symbolKey: Symbol(123)
 };
 
 console.log(`exampleObj is defined - ${t(exampleObj).isDefined}`); // true
@@ -42,12 +43,31 @@ const arr = [];
 console.log(`arr is an Array - ${t(arr).isArray}`); // true
 console.log(`arr is an empty Array - ${t(arr).isEmptyArray}`); // true
 
+const date = new Date();
+console.log(`type of 'date' is Date - ${t(date).isDate}`); // true
+console.log(`type of 'obj' is not date - ${t(obj).isDate}`); // false
+console.log(`type of 'arr' is not date - ${t(arr).isDate}`); // false
+
+const exampleDateObj = {
+  key: {
+    nestedGoodDate: 'new Date()'
+  }
+};
+console.log(`exampleDateObj.key.nestedGoodDate is defined - ${t(exampleDateObj, 'key.nestedGoodDate').isDefined}`); // true
+console.log(`exampleDateObj.key.nestedGoodDate is of type Date - ${t(exampleDateObj, 'key.nestedGoodDate').isDate}`); // true
+
+const mySymbol = Symbol('symbol-description');
+console.log(`is 'mySymbol' a Symbol - ${t(mySymbol).isSymbol}`);
+console.log(`is 'obj' a Symbol? - ${t(obj).isSymbol}`);
+console.log(`is 'arr' a Symbol? - ${t(arr).isSymbol}`);
+
 // Schema Checks
 const validSchema = {
   goodKey: Schema.String,
   nestedKey: {
     nestedGoodKey: Schema.String
-  }
+  },
+  symbolKey: Schema.Symbol
 };
 console.log(`exampleObj matches validSchema - ${t(exampleObj, validSchema).isValid}`); // true