+++ index.d.ts
@@ -39,7 +39,9 @@ declare module "typy" {
 		readonly isNumber: boolean
 		readonly isArray: boolean
 		readonly isEmptyArray: boolean
-		readonly isFunction: boolean
+    readonly isFunction: boolean
+		readonly isDate: boolean
+		readonly isSymbol: boolean
 		readonly safeObject: any
 		readonly safeObjectOrEmpty: any
 		readonly safeString: string