+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## To Be Released
 
+* New APIs added: 
+  * `t(dateObj).isDate` - is available.
+  * `t(Symbol('desc')).isSymbol` - is available
+
 ### version (undecided)
 
 * what's next